Reconfigure timers again

+++ Inc/pwm_fan.h
@@ -22,6 +22,6 @@ typedef struct {
 
 void pwm_fan_init(PWM_Fan_HandleTypeDef *fan, TIM_HandleTypeDef *htim_pwm, TIM_HandleTypeDef *htim_tacho, uint32_t pwm_channel, uint32_t tacho_channel);
 void pwm_fan_set(PWM_Fan_HandleTypeDef *fan, double target_speed);
-float pwm_fan_update(PWM_Fan_HandleTypeDef *fan);
+double pwm_fan_update(PWM_Fan_HandleTypeDef *fan);
 
 #endif //PWM_FAN_H
